requirements.txt
django
djangorestframework
channels 
drf-yasg 
redis
psycopg2-binary
